+++ service/mongodb/controller.go
@@ -3,6 +3,7 @@ package mongodb
 import (
 	"context"
 	"fmt"
+	"time"
 
 	pb "github.com/meateam/permission-service/proto"
 	"github.com/meateam/permission-service/service"
@@ -35,8 +36,9 @@ func (c Controller) CreatePermission(
 	userID string,
 	role pb.Role,
 	creator string,
-	override bool) (service.Permission, error) {
-	permission := &BSON{FileID: fileID, UserID: userID, Role: role, Creator: creator}
+	override bool,
+	appID string) (service.Permission, error) {
+	permission := &BSON{FileID: fileID, UserID: userID, Role: role, Creator: creator, AppID: appID, CreatedAt: time.Now(), UpdatedAt: time.Now()}
 	createdPermission, err := c.store.Create(ctx, permission, override)
 	if err != nil {
 		return nil, fmt.Errorf("failed creating permission: %v", err)
@@ -163,19 +165,59 @@ func (c Controller) GetFilePermissions(ctx context.Context,
 // otherwise returns nil and any error if occurred.
 func (c Controller) GetUserPermissions(
 	ctx context.Context,
-	userID string) ([]*pb.GetUserPermissionsResponse_FileRole, error) {
-	filter := bson.D{
+	userID string, pageNum int64, pageSize int64, isShared bool, appID string) (*pb.GetUserPermissionsResponse, error) {
+
+	// Check if one is negative and the other is not
+	if pageNum < 0 || pageSize < 0 {
+		return nil, fmt.Errorf("pageNum %d and pageSize %d must both be non-negative", pageNum, pageSize)
+	}
+
+	var filter bson.D
+
+	filter = append(filter, bson.E{
+		Key:   PermissionBSONUserIDField,
+		Value: userID,
+	})
+
+	if isShared {
+		filter = append(filter, bson.E{
+			Key:   PermissionBSONCreatorField,
+			Value: bson.M{"$ne": userID},
+		})
+	}
+
+	if appID != "" {
+		filter = append(filter, bson.E{
+			Key:   PermissionBSONAppIDField,
+			Value: appID,
+		})
+	}
+
+	// Sort by decreasing createdAt order (most recent to oldest).
+	sort := bson.D{
 		bson.E{
-			Key:   PermissionBSONUserIDField,
-			Value: userID,
+			Key:   PermissionBSONUpdatedAtField,
+			Value: -1,
 		},
 	}
 
-	permissions, err := c.store.GetAll(ctx, filter)
+	// Get permissions by page, sorted by reverse mongoID
+	pageRes, err := c.store.GetUserPermissionsByPage(ctx, pageNum, pageSize, sort, filter)
 	if err != nil {
 		return nil, err
 	}
 
+	// Go over the page of permissions received and reformat them
+	filePermissions := c.reformatFilePermissions(pageRes.permissions)
+
+	return &pb.GetUserPermissionsResponse{Permissions: filePermissions, ItemCount: pageRes.itemCount, PageNum: pageRes.pageNum}, nil
+
+}
+
+// reformatFilePermissions receives an array of service.Permission and
+// returns them as an array of *pb.GetUserPermissionsResponse_FileRole,
+// while keeping the order in which they were received
+func (c Controller) reformatFilePermissions(permissions []service.Permission) []*pb.GetUserPermissionsResponse_FileRole {
 	filePermissions := make([]*pb.GetUserPermissionsResponse_FileRole, 0, len(permissions))
 	for _, permission := range permissions {
 		filePermissions = append(filePermissions, &pb.GetUserPermissionsResponse_FileRole{
@@ -185,7 +227,7 @@ func (c Controller) GetUserPermissions(
 		})
 	}
 
-	return filePermissions, nil
+	return filePermissions
 }
 
 // DeleteFilePermissions deletes all permissions that exist for fileID and