+++ service/mongodb/mongodb.go
@@ -2,6 +2,7 @@ package mongodb
 
 import (
 	"fmt"
+	"time"
 
 	pb "github.com/meateam/permission-service/proto"
 	"go.mongodb.org/mongo-driver/bson/primitive"
@@ -9,11 +10,14 @@ import (
 
 // BSON is the structure that represents a permission as it's stored.
 type BSON struct {
-	ID      primitive.ObjectID `bson:"_id,omitempty"`
-	FileID  string             `bson:"fileID,omitempty"`
-	UserID  string             `bson:"userID,omitempty"`
-	Role    pb.Role            `bson:"role"`
-	Creator string             `bson:"creator"`
+	ID        primitive.ObjectID `bson:"_id,omitempty"`
+	FileID    string             `bson:"fileID,omitempty"`
+	UserID    string             `bson:"userID,omitempty"`
+	Role      pb.Role            `bson:"role"`
+	Creator   string             `bson:"creator"`
+	AppID     string             `bson:"appID"`
+	CreatedAt time.Time          `bson:"createdAt" json:"createdAt,omitempty"`
+	UpdatedAt time.Time          `bson:"updatedAt" json:"updatedAt,omitempty"`
 }
 
 // GetID returns the string value of the b.ID.
@@ -116,6 +120,35 @@ func (b *BSON) SetCreator(creator string) error {
 	return nil
 }
 
+// GetAppID returns b.AppID.
+func (b BSON) GetAppID() string {
+	return b.AppID
+}
+
+// SetAppID sets b.AppID to appID.
+func (b *BSON) SetAppID(appID string) error {
+	if b == nil {
+		panic("b == nil")
+	}
+
+	if appID == "" {
+		return fmt.Errorf("appID is required")
+	}
+
+	b.AppID = appID
+	return nil
+}
+
+// GetCreatedAt returns b.CreatedAt.
+func (b BSON) GetCreatedAt() time.Time {
+	return b.CreatedAt
+}
+
+// GetUpdatedAt returns b.UpdatedAt.
+func (b BSON) GetUpdatedAt() time.Time {
+	return b.UpdatedAt
+}
+
 // MarshalProto marshals b into a permission.
 func (b BSON) MarshalProto(permission *pb.PermissionObject) error {
 	permission.Id = b.GetID()
@@ -123,6 +156,9 @@ func (b BSON) MarshalProto(permission *pb.PermissionObject) error {
 	permission.UserID = b.GetUserID()
 	permission.Role = b.GetRole()
 	permission.Creator = b.GetCreator()
+	permission.AppID = b.GetAppID()
+	permission.CreatedAt = b.GetCreatedAt().Unix()
+	permission.UpdatedAt = b.GetUpdatedAt().Unix()
 
 	return nil
 }