+++ service/mongodb/store.go
@@ -30,11 +30,27 @@ const (
 
 	// PermissionBSONCreatorField is the name of the creator field in BSON.
 	PermissionBSONCreatorField = "creator"
+
+	// PermissionBSONAppIDField is the name of the appID field in BSON.
+	PermissionBSONAppIDField = "appID"
+
+	// PermissionBSONCreatedAtField is the name of the createdAt field in BSON.
+	PermissionBSONCreatedAtField = "createdAt"
+
+	// PermissionBSONUpdatedAtField is the name of the updatedAt field in BSON.
+	PermissionBSONUpdatedAtField = "updatedAt"
 )
 
 // MongoStore holds the mongodb database and implements Store interface.
 type MongoStore struct {
 	DB *mongo.Database
+}
+
+// PagingRes is the struct of the result from the pagination requests
+type PagingRes struct {
+	permissions []service.Permission
+	itemCount   int64
+	pageNum     int64
 }
 
 // newMongoStore returns a new store.
@@ -103,6 +119,21 @@ func (s MongoStore) Create(
 		return nil, fmt.Errorf("creator is required")
 	}
 
+	appID := permission.GetAppID()
+	if appID == "" {
+		return nil, fmt.Errorf("appID is required")
+	}
+
+	createdAt := permission.GetCreatedAt()
+	if appID == "" {
+		return nil, fmt.Errorf("current time could not be obtained")
+	}
+
+	updatedAt := permission.GetUpdatedAt()
+	if appID == "" {
+		return nil, fmt.Errorf("current time could not be obtained")
+	}
+
 	filter := bson.D{
 		bson.E{
 			Key:   PermissionBSONFileIDField,
@@ -112,6 +143,15 @@ func (s MongoStore) Create(
 			Key:   PermissionBSONUserIDField,
 			Value: userID,
 		},
+	}
+
+	// Check existance of the permission. If the error received is not `404` then abort.
+	// In case the permission exists, do not update createdAt.
+	existingPermission, existingPermissionErr := s.Get(ctx, filter)
+	if existingPermissionErr != nil && existingPermissionErr != mongo.ErrNoDocuments {
+		return nil, existingPermissionErr
+	} else if existingPermissionErr == nil {
+		createdAt = existingPermission.GetCreatedAt()
 	}
 
 	newPermission := bson.D{
@@ -131,6 +171,18 @@ func (s MongoStore) Create(
 			Key:   PermissionBSONCreatorField,
 			Value: creator,
 		},
+		bson.E{
+			Key:   PermissionBSONAppIDField,
+			Value: appID,
+		},
+		bson.E{
+			Key:   PermissionBSONCreatedAtField,
+			Value: createdAt,
+		},
+		bson.E{
+			Key:   PermissionBSONUpdatedAtField,
+			Value: updatedAt,
+		},
 	}
 
 	update := bson.D{
@@ -140,16 +192,9 @@ func (s MongoStore) Create(
 		},
 	}
 
-	// In case override is false, check if there is a permission, and if there is one, return it.
-	if !override {
-		existingPermission, err := s.Get(ctx, filter)
-		if err != nil && err != mongo.ErrNoDocuments {
-			return nil, err
-		}
-
-		if err == nil {
-			return existingPermission, nil
-		}
+	// In case override is false, check if there is a permission, and if there is one - return it.
+	if !override && existingPermissionErr == nil {
+		return existingPermission, nil
 	}
 
 	// If override is true, or false and there is no permission existing,
@@ -174,7 +219,8 @@ func (s MongoStore) Get(ctx context.Context, filter interface{}) (service.Permis
 	collection := s.DB.Collection(PermissionCollectionName)
 
 	permission := &BSON{}
-	err := collection.FindOne(ctx, filter).Decode(permission)
+	foundPerm := collection.FindOne(ctx, filter)
+	err := foundPerm.Decode(permission)
 	if err != nil {
 		return nil, err
 	}
@@ -210,6 +256,47 @@ func (s MongoStore) GetAll(ctx context.Context, filter interface{}) ([]service.P
 	}
 
 	return permissions, nil
+}
+
+// GetUserPermissionsByPage returns a slice of the permissions requested by the filter,
+// in the page they belong to by page number and page size.
+// sortBy is the field by which the sorting of the permissions will be committed,
+// defaulting to reverse mongoID.
+func (s MongoStore) GetUserPermissionsByPage(ctx context.Context, pn int64, ps int64, sortBy bson.D, filter interface{}) (*PagingRes, error) {
+	collection := s.DB.Collection(PermissionCollectionName)
+
+	itemCount, err := collection.CountDocuments(ctx, filter)
+	if err != nil {
+		return nil, err
+	}
+
+	opts := options.Find().SetSort(sortBy).SetLimit(ps).SetSkip(pn * ps)
+
+	cur, err := collection.Find(ctx, filter, opts)
+	if err != nil {
+		return nil, err
+	}
+
+	defer cur.Close(ctx)
+
+	permissions := []service.Permission{}
+	for cur.Next(ctx) {
+		permission := &BSON{}
+		err := cur.Decode(permission)
+		if err != nil {
+			return nil, err
+		}
+
+		permissions = append(permissions, permission)
+	}
+
+	if err := cur.Err(); err != nil {
+		return nil, err
+	}
+
+	var pr *PagingRes = &PagingRes{permissions: permissions, pageNum: pn, itemCount: itemCount}
+
+	return pr, nil
 }
 
 // Delete finds the first permission that matches filter and deletes it,