+++ CHANGELOG.md
@@ -11,8 +11,21 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - ISSUE([99](https://github.com/meateam/drive-project/issues/99)): update pkg, delete unused pkg and update deps
 
 
+## [v3.1.1] - 2021-01-19
+
+### Fixed
+
+- hotfix: ([21](https://github.com/meateam/permission-service/pull/21)): changed updated_at, created_at to updatedAt, createdAt
+
 ### Added
 
+- minor: ([21](https://github.com/meateam/permission-service/pull/21)): added updatedAt and createdAt fields, pagination sorting defaults to updated_at field
+
+## [v3.0.0] - 2021-01-13
+
+### Added
+
+- major: ([10](https://github.com/meateam/permission-service/pull/10)): added pagination, and appID in create
 - FEAT([17](https://github.com/meateam/permission-service/pull/17)): add ci and test file
 
 ### Refactor
@@ -26,4 +39,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - FEAT([9](https://github.com/meateam/permission-service/pull/9)): RPC method GetPermissionByMongoID
 
 [unreleased]: https://github.com/meateam/permission-service/compare/master...develop
+[v3.1.1]: https://github.com/meateam/permission-service/compare/v3.0.0...v3.1.1
+[v3.0.0]: https://github.com/meateam/permission-service/compare/v2.0.0...v3.0.0
 [v2.0.0]: https://github.com/meateam/permission-service/compare/v1.3...v2.0.0